+++ components/index.js
@@ -1,2 +1,4 @@
 export { Account } from './Account'
 export { Connect } from './Connect'
+export { TokenDisplay } from './TokenDisplay'
+export { Erc1155Balance } from './Erc1155Balance'
