+++ hooks/index.js
@@ -1,1 +1,3 @@
 export { useMounted } from './useMounted'
+export { useTokenURI } from './useTokenUri'
+export { useHas1155Token } from './has1155Token'
