+++ hooks/useTokenUri.js
@@ -1,20 +1,40 @@
 import { useContractRead } from 'wagmi'
 import erc1155Contract from '../utils/abis/erc1155Contract.json'
 
-export const useTemplateURI = (contract) => {
+export const useTokenURI = (contract, tokenId) => {
   const {
-    data: templateURI,
+    data: tokenURI,
     isError,
     error
   } = useContractRead({
     addressOrName: contract,
     contractInterface: erc1155Contract,
-    functionName: 'templateURI'
+    functionName: 'uri',
+    args: [tokenId]
   })
-
+  const processedTokenURI = processTokenURI(
+    typeof tokenURI == 'string' ? tokenURI : '',
+    tokenId
+  )
   return {
-    templateURI,
-    templateUriIsError: isError,
-    templateUriError: error
+    processedTokenURI,
+    tokenURI,
+    tokenUriIsError: isError,
+    tokenUriError: error
   }
+}
+
+// OpenSea does not return an URI for the token's metadata, but a pattern
+// We need to replace the ID in '0x{id}' with the token ID in hex
+const processTokenURI = (tokenURI, tokenId) => {
+  // Do not alter the URI if it is not an OpenSea API one
+  if (tokenURI.indexOf('api.opensea.io/api') == -1) return tokenURI
+
+  // Do not alter the URI if it does not come with '0x{id}' at the end
+  const tokenIdIndex = tokenURI.indexOf('0x{id}')
+  if (tokenIdIndex == -1) return tokenURI
+
+  const processedTokenURI = tokenURI.substring(0, tokenIdIndex) + '0x'
+  const tokenIdHex = BigInt(tokenId).toString(16)
+  return processedTokenURI + tokenIdHex + '?format=json'
 }
