+++ pages/index.js
@@ -1,8 +1,8 @@
 import { Box, Container, Heading } from '@chakra-ui/react'
 import { useAccount } from 'wagmi'
-import { Account, Connect } from '../components'
+import { Connect } from '../components'
 import { useMounted } from '../hooks'
-import { Erc1155Balance } from '../components/erc1155Balance'
+import { Erc1155Balance } from '../components'
 
 export default function Home() {
   const isMounted = useMounted()
