+++ utils/index.js
@@ -1,0 +1,1 @@
+export { contractAddress } from './contractAddress'
